Improve MIME to file extension function by including more MIME types
